Add files via upload
A series of completed projects from coursework at TX State University. The titles are self descriptive of the project. The ALU design was part of a project for VLSI and included the use of System Verilog, GitHub repository, and Visual Studio IDE with use of test bench. The file opens in Visual Studio as long as the user has the software installed.