Add files via upload
Capstone Senior Design Project is a two semester course. This is the work from the first of those two semesters (currently in the second semester for Spring '21). The first semester was the research and design phase; which consisted of the use of 4NEC2 software to design a narrow bandwidth HF antenna capable of multiband reception of HAM radio bandwidths. The second semester consists of the construction and test phase culminating in the showcase of the project at senior design day.